# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/roles/my/channel/{channelId}": {
      "get": {
        "operationId": "RolesController_getMyRolesForChannel",
        "parameters": [
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRolesResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Roles"
        ]
      }
    },
...
    "/api/roles/user/{userId}/channel/{channelId}": {
      "get": {
        "operationId": "RolesController_getUserRolesForChannel",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRolesResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Roles"
        ]
      }
    },
...
    "/api/channels": {
      "post": {
        "operationId": "ChannelsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChannelDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelDto"
                }
              }
            }
          }
        },
        "tags": [
          "Channels"
        ]
      }
    },
    "/api/channels/community/{communityId}": {
      "get": {
        "operationId": "ChannelsController_findAllForCommunity",
        "parameters": [
          {
            "name": "communityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChannelDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Channels"
        ]
      }
    },
    "/api/channels/community/{communityId}/mentionable": {
      "get": {
        "operationId": "ChannelsController_getMentionableChannels",
        "parameters": [
          {
            "name": "communityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChannelDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Channels"
        ]
      }
    },
    "/api/channels/{id}": {
      "get": {
        "operationId": "ChannelsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelDto"
                }
              }
            }
          }
        },
        "tags": [
          "Channels"
        ]
      },
      "patch": {
        "operationId": "ChannelsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChannelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelDto"
                }
              }
            }
          }
        },
        "tags": [
          "Channels"
        ]
      },
      "delete": {
        "operationId": "ChannelsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Channels"
        ]
      }
    },
    "/api/channels/{id}/move-up": {
      "post": {
        "operationId": "ChannelsController_moveUp",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveChannelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChannelDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Channels"
        ]
      }
    },
    "/api/channels/{id}/move-down": {
      "post": {
        "operationId": "ChannelsController_moveDown",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveChannelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChannelDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Channels"
        ]
      }
    },
...
    "/api/community/{id}": {
      "get": {
        "operationId": "CommunityController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Community"
        ]
      },
      "patch": {
        "operationId": "CommunityController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommunityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Community"
        ]
      },
      "delete": {
        "operationId": "CommunityController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Community"
        ]
      }
    },
...
    "/api/messages/channel/{channelId}": {
      "get": {
        "operationId": "MessagesController_findAllForChannel",
        "parameters": [
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "continuationToken",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedMessagesResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Messages"
        ]
      }
    },
    "/api/messages/search/channel/{channelId}": {
      "get": {
        "operationId": "MessagesController_searchChannelMessages",
        "parameters": [
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnrichedMessageDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Messages"
        ]
      }
    },
...
    "/api/channel-membership/channel/{channelId}": {
      "get": {
        "operationId": "ChannelMembershipController_findAllForChannel",
        "parameters": [
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChannelMembershipResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "ChannelMembership"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kraken API",
    "description": "Kraken Chat Application API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UserEntity": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "USER"
            ]
          },
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "bannerUrl": {
            "type": "string",
            "nullable": true
          },
          "lastSeen": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "role",
          "id",
          "username",
          "avatarUrl",
          "bannerUrl",
          "lastSeen",
          "displayName",
          "bio",
          "status"
        ]
      },
...
      "RoleDto": {
        "type": "object",
        "properties": {
          "actions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "DELETE_MESSAGE",
                "DELETE_CHANNEL",
                "DELETE_COMMUNITY",
                "DELETE_INVITE",
                "DELETE_USER",
                "DELETE_ROLE",
                "DELETE_ALIAS_GROUP",
                "DELETE_ALIAS_GROUP_MEMBER",
                "DELETE_INSTANCE_INVITE",
                "DELETE_MEMBER",
                "DELETE_REACTION",
                "CREATE_MESSAGE",
                "CREATE_CHANNEL",
                "CREATE_COMMUNITY",
                "CREATE_INVITE",
                "CREATE_USER",
                "CREATE_ROLE",
                "CREATE_ALIAS_GROUP",
                "CREATE_ALIAS_GROUP_MEMBER",
                "CREATE_INSTANCE_INVITE",
                "CREATE_MEMBER",
                "CREATE_REACTION",
                "JOIN_CHANNEL",
                "READ_MESSAGE",
                "READ_CHANNEL",
                "READ_COMMUNITY",
                "READ_ALL_COMMUNITIES",
                "READ_USER",
                "READ_ROLE",
                "READ_ALIAS_GROUP",
                "READ_ALIAS_GROUP_MEMBER",
                "READ_INSTANCE_INVITE",
                "READ_MEMBER",
                "UPDATE_COMMUNITY",
                "UPDATE_CHANNEL",
                "UPDATE_MESSAGE",
                "UPDATE_USER",
                "UPDATE_ROLE",
                "UPDATE_ALIAS_GROUP",
                "UPDATE_ALIAS_GROUP_MEMBER",
                "UPDATE_INSTANCE_INVITE",
                "UPDATE_MEMBER",
                "CAPTURE_REPLAY",
                "READ_INSTANCE_SETTINGS",
                "UPDATE_INSTANCE_SETTINGS",
                "READ_INSTANCE_STATS",
                "MANAGE_USER_STORAGE",
                "BAN_USER",
                "KICK_USER",
                "TIMEOUT_USER",
                "UNBAN_USER",
                "PIN_MESSAGE",
                "UNPIN_MESSAGE",
                "DELETE_ANY_MESSAGE",
                "VIEW_BAN_LIST",
                "VIEW_MODERATION_LOGS"
              ]
            }
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "isDefault": {
            "type": "boolean"
          }
        },
        "required": [
          "actions",
          "id",
          "name",
          "createdAt",
          "isDefault"
        ]
      },
      "UserRolesResponseDto": {
        "type": "object",
        "properties": {
          "resourceType": {
            "type": "string",
            "enum": [
              "COMMUNITY",
              "CHANNEL",
              "INSTANCE"
            ],
            "nullable": true
          },
          "userId": {
            "type": "string"
          },
          "resourceId": {
            "type": "string",
            "nullable": true
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleDto"
            }
          }
        },
        "required": [
          "resourceType",
          "userId",
          "resourceId",
          "roles"
        ]
      },
...
      "CreateChannelDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "TEXT",
              "VOICE"
            ]
          },
          "isPrivate": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "communityId": {
            "type": "string"
          },
          "position": {
            "type": "number"
          },
          "slowmodeSeconds": {
            "type": "number",
            "minimum": 0,
            "maximum": 21600
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "type",
          "isPrivate",
          "name",
          "communityId",
          "position",
          "slowmodeSeconds",
          "id",
          "createdAt"
        ]
      },
      "ChannelDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "TEXT",
              "VOICE"
            ]
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "communityId": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "position": {
            "type": "number"
          },
          "slowmodeSeconds": {
            "type": "number"
          },
          "isPrivate": {
            "type": "boolean"
          }
        },
        "required": [
          "type",
          "id",
          "name",
          "communityId",
          "createdAt",
          "position",
          "slowmodeSeconds",
          "isPrivate"
        ]
      },
      "UpdateChannelDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "TEXT",
              "VOICE"
            ]
          },
          "isPrivate": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "communityId": {
            "type": "string"
          },
          "position": {
            "type": "number"
          },
          "slowmodeSeconds": {
            "type": "number",
            "minimum": 0,
            "maximum": 21600
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        }
      },
      "MoveChannelDto": {
        "type": "object",
        "properties": {
          "communityId": {
            "type": "string"
          }
        },
        "required": [
          "communityId"
        ]
      },
...
      "CommunityResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "avatar": {
            "type": "string",
            "nullable": true
          },
          "banner": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "avatar",
          "banner",
          "createdAt"
        ]
      },
      "UpdateCommunityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 500
          },
          "avatar": {
            "type": "string",
            "nullable": true
          },
          "banner": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "SpanDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "PLAINTEXT",
              "USER_MENTION",
              "SPECIAL_MENTION",
              "COMMUNITY_MENTION",
              "ALIAS_MENTION"
            ]
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "specialKind": {
            "type": "string",
            "nullable": true
          },
          "communityId": {
            "type": "string",
            "nullable": true
          },
          "aliasId": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "type",
          "text",
          "userId",
          "specialKind",
          "communityId",
          "aliasId"
        ]
      },
      "EnrichedAttachment": {
        "type": "object",
        "properties": {
          "fileType": {
            "type": "string",
            "enum": [
              "IMAGE",
              "VIDEO",
              "AUDIO",
              "DOCUMENT",
              "OTHER"
            ]
          },
          "id": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "size": {
            "type": "number"
          }
        },
        "required": [
          "fileType",
          "id",
          "filename",
          "mimeType",
          "size"
        ]
      },
      "ReactionDto": {
        "type": "object",
        "properties": {
          "emoji": {
            "type": "string"
          },
          "userIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "emoji",
          "userIds"
        ]
      },
      "EnrichedMessageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "channelId": {
            "type": "string",
            "nullable": true
          },
          "directMessageGroupId": {
            "type": "string",
            "nullable": true
          },
          "authorId": {
            "type": "string"
          },
          "spans": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpanDto"
            }
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EnrichedAttachment"
            }
          },
          "pendingAttachments": {
            "type": "number",
            "nullable": true
          },
          "reactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReactionDto"
            }
          },
          "replyCount": {
            "type": "number"
          },
          "lastReplyAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "pinned": {
            "type": "boolean"
          },
          "pinnedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "pinnedBy": {
            "type": "string",
            "nullable": true
          },
          "sentAt": {
            "format": "date-time",
            "type": "string"
          },
          "editedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "channelId",
          "directMessageGroupId",
          "authorId",
          "spans",
          "attachments",
          "pendingAttachments",
          "reactions",
          "replyCount",
          "lastReplyAt",
          "pinned",
          "pinnedAt",
          "pinnedBy",
          "sentAt",
          "editedAt",
          "deletedAt"
        ]
      },
      "PaginatedMessagesResponseDto": {
        "type": "object",
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EnrichedMessageDto"
            }
          },
          "continuationToken": {
            "type": "string"
          }
        },
        "required": [
          "messages"
        ]
      },
...
      "ChannelMembershipResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserEntity"
          },
          "channelId": {
            "type": "string"
          },
          "joinedAt": {
            "format": "date-time",
            "type": "string"
          },
          "addedBy": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "channelId",
          "joinedAt"
        ]
      },
...
    }
  }
}